requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
discord.py==2.0.1
frozenlist==1.3.1
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
idna==3.4
msgpack==1.0.4
multidict==6.0.2
pip==22.3
protobuf==4.21.9
proto-plus==1.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
requests==2.28.1
rsa==4.9
setuptools==65.5.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
yarl==1.8.1
